feat(settings): move Stats for Nerds into a sheet with Copy as Text
The disclosure group pushed forty rows of diagnostics into the Output tab and
made it scroll past the controls people actually use. The stats now open in a
sheet with one section per topic, in the style of the Wi-Fi details panel in
System Settings. Copy as Text puts every section on the clipboard, which is
the first step toward the support report on the v1 list.

The label formatting moved out of the view into OutputDiagnosticsReport so
the sheet, the summary card, and the text export share it. Output tab
buttons use title case like the rest of the window.
