+++ Sources/GlassEQSettingsUI/SettingsView.swift
@@ -40,7 +40,7 @@ func localized(_ value: String.LocalizationValue) -> String {
     String(localized: value, bundle: settingsResourcesBundle)
 }
 
-private func localizedDecimal(
+func localizedDecimal(
     _ value: Double,
     minimumFractionDigits: Int,
     maximumFractionDigits: Int,
@@ -137,19 +137,19 @@ func playbackFramesToMilliseconds(
     return frames / sampleRate * 1_000
 }
 
-private func localizedInteger(_ value: Int) -> String {
+func localizedInteger(_ value: Int) -> String {
     value.formatted(.number.locale(.autoupdatingCurrent))
 }
 
-private func localizedInteger(_ value: UInt32) -> String {
+func localizedInteger(_ value: UInt32) -> String {
     UInt64(value).formatted(.number.locale(.autoupdatingCurrent))
 }
 
-private func localizedInteger(_ value: UInt64) -> String {
+func localizedInteger(_ value: UInt64) -> String {
     value.formatted(.number.locale(.autoupdatingCurrent))
 }
 
-private func localizedDecibels(_ value: Double, fractionDigits: Int = 1) -> String {
+func localizedDecibels(_ value: Double, fractionDigits: Int = 1) -> String {
     let number = localizedDecimal(
         value,
         minimumFractionDigits: fractionDigits,
@@ -159,7 +159,7 @@ private func localizedDecibels(_ value: Double, fractionDigits: Int = 1) -> Stri
     return localized("\(number) dB")
 }
 
-private func localizedFrequency(_ value: Double) -> String {
+func localizedFrequency(_ value: Double) -> String {
     if value >= 1_000 {
         let number = localizedDecimal(value / 1_000, minimumFractionDigits: 1, maximumFractionDigits: 1)
         return localized("\(number) kHz")
@@ -168,12 +168,12 @@ private func localizedFrequency(_ value: Double) -> String {
     return localized("\(number) Hz")
 }
 
-private func localizedFrameCount(_ value: Int) -> String {
+func localizedFrameCount(_ value: Int) -> String {
     let number = localizedInteger(value)
     return value == 1 ? localized("\(number) frame") : localized("\(number) frames")
 }
 
-private func localizedFrameCount(_ value: UInt32) -> String {
+func localizedFrameCount(_ value: UInt32) -> String {
     let number = localizedInteger(value)
     return value == 1 ? localized("\(number) frame") : localized("\(number) frames")
 }
@@ -221,7 +221,7 @@ func settingsSnapshotAfterCommand(
     return latest
 }
 
-private func localizedLatency(milliseconds: Double) -> String {
+func localizedLatency(milliseconds: Double) -> String {
     let number = localizedDecimal(milliseconds, minimumFractionDigits: 2, maximumFractionDigits: 2)
     return localized("\(number) ms")
 }
@@ -2907,7 +2907,7 @@ private struct OutputTab: View {
     var onRetryAutomaticAggregateBuffer: () -> Void
     var onRetryAudioEngine: () -> Void
     var onOpenPrivacySettings: () -> Void
-    @State private var showsNerdStats = false
+    @State private var isShowingDiagnostics = false
 
     var body: some View {
         HStack(alignment: .top, spacing: 16) {
@@ -2947,13 +2947,13 @@ private struct OutputTab: View {
 
                     if snapshot.aggregateBuffer.mode == .automatic,
                        snapshot.aggregateBuffer.automaticFrameSize > 16 {
-                        Button(localized("Retry 16 frames")) {
+                        Button(localized("Retry 16 Frames")) {
                             onRetryAutomaticAggregateBuffer()
                         }
                         .controlSize(.large)
                     } else if let fixedFrameSize,
                               snapshot.currentOutputBufferFrameSize > fixedFrameSize {
-                        Button(localized("Retry \(fixedFrameSize) frames")) {
+                        Button(localized("Retry \(fixedFrameSize) Frames")) {
                             onSetAggregateBufferMode(snapshot.aggregateBuffer.mode)
                         }
                         .controlSize(.large)
@@ -2973,7 +2973,7 @@ private struct OutputTab: View {
                         .disabled(isProfileStoreProtected || snapshot.currentOutputUID.isEmpty)
                         .controlSize(.large)
 
-                        Button(localized("Set as fallback profile")) {
+                        Button(localized("Set as Fallback")) {
                             onSetFallback()
                         }
                         .disabled(isProfileStoreProtected)
@@ -2994,7 +2994,7 @@ private struct OutputTab: View {
                     LabeledContent(localized("Current Output"), value: snapshot.currentOutputName)
                     LabeledContent(localized("Active Profile"), value: snapshot.activeProfileName)
                     LabeledContent(localized("Buffer"), value: bufferSummary)
-                    LabeledContent(localized("Added Latency"), value: addedLatencyLabel)
+                    LabeledContent(localized("Added Latency"), value: report.addedLatencyLabel)
                     LabeledContent(
                         localized("Underrun Events"),
                         value: snapshot.metrics.playbackUnderrunEvents == 0
@@ -3006,7 +3006,7 @@ private struct OutputTab: View {
                         .foregroundStyle(.secondary)
                         .fixedSize(horizontal: false, vertical: true)
                     HStack {
-                        Button(localized("Retry audio engine")) {
+                        Button(localized("Retry Audio Engine")) {
                             onRetryAudioEngine()
                         }
                         .controlSize(.large)
@@ -3020,98 +3020,29 @@ private struct OutputTab: View {
                 .frame(maxWidth: .infinity, alignment: .topLeading)
                 .cardPanel(padding: 16)
 
-                VStack(alignment: .leading, spacing: 8) {
-                    DisclosureGroup(isExpanded: $showsNerdStats) {
-                        VStack(alignment: .leading, spacing: 10) {
-                            diagnosticSectionTitle(localized("Observation"))
-                            LabeledContent(localized("Reset"), value: diagnosticsResetLabel)
-                            LabeledContent(localized("Observed"), value: observationDurationLabel)
-                            LabeledContent(localized("Current Runtime"), value: runtimeDurationLabel)
-
-                            Divider()
-                            diagnosticSectionTitle(localized("Timing"))
-                            Text(localized("p50 / p99 / p99.9 / p99.99 / max"))
-                                .font(.caption.monospacedDigit())
-                                .foregroundStyle(.secondary)
-                            timingRows
-
-                            Divider()
-                            diagnosticSectionTitle(localized("Reliability"))
-                            LabeledContent(localized("Captured Frames"), value: localizedInteger(snapshot.metrics.capturedFrames))
-                            LabeledContent(localized("Played Frames"), value: localizedInteger(snapshot.metrics.playedFrames))
-                            LabeledContent(localized("Underrun Events / Frames"), value: underrunDetailLabel)
-                            LabeledContent(localized("Dropped Input / Buffered"), value: droppedFramesLabel)
-                            LabeledContent(localized("Saturated Samples"), value: localizedInteger(snapshot.metrics.saturatedSamples))
-                            LabeledContent(localized("Deadline Misses"), value: deadlineMissesLabel)
-                            LabeledContent(localized("Discontinuities"), value: discontinuityLabel)
-                            if usesSeparateClockDiagnostics {
-                                LabeledContent(localized("Ring Gate Failures"), value: localizedInteger(snapshot.metrics.ringGateContentionFailures))
-                                LabeledContent(localized("Buffered / Peak"), value: bufferedFramesLabel)
-                                LabeledContent(localized("Clock Correction"), value: playbackRateCorrectionLabel)
-                                LabeledContent(localized("Servo Buffer"), value: servoBufferLabel)
-                                LabeledContent(localized("Bridge Latency"), value: bridgeLatencyLabel)
-                                LabeledContent(localized("Bridge Latency Range"), value: bridgeLatencyRangeLabel)
-                            }
-
-                            Divider()
-                            diagnosticSectionTitle(localized("Recovery"))
-                            LabeledContent(localized("Runtime Rebuilds"), value: localizedInteger(diagnostics.recovery.runtimeRebuilds))
-                            LabeledContent(localized("Automatic Recoveries"), value: localizedInteger(diagnostics.recovery.automaticRecoveries))
-                            LabeledContent(localized("Buffer Escalations"), value: localizedInteger(diagnostics.recovery.bufferEscalations))
-                            LabeledContent(localized("Headset Fallbacks"), value: localizedInteger(diagnostics.recovery.headsetFallbacks))
-                            LabeledContent(localized("Last Recovery"), value: lastRecoveryLabel)
-
-                            Divider()
-                            diagnosticSectionTitle(localized("Callback Sizes"))
-                            LabeledContent(localized("Capture"), value: callbackSizeHistogramLabel(snapshot.metrics.captureCallbackSizeObservations))
-                            LabeledContent(localized("Output"), value: callbackSizeHistogramLabel(snapshot.metrics.playbackCallbackSizeObservations))
-                            LabeledContent(localized("Capture / Output Peak"), value: callbackPeaksLabel)
-
-                            Divider()
-                            diagnosticSectionTitle(localized("Timestamps"))
-                            LabeledContent(localized("Last Sample-Time Jumps"), value: timestampJumpLabel)
-                            LabeledContent(localized("Last Host-Time Errors"), value: hostIntervalErrorLabel)
-                            LabeledContent(localized("Jump Interval min / avg / max"), value: timestampJumpIntervalLabel)
-                            LabeledContent(localized("Input Age min / avg / max"), value: inputAgeLabel)
-                            LabeledContent(localized("Output Lead min / avg / max"), value: outputLeadLabel)
-
-                            Divider()
-                            diagnosticSectionTitle(localized("Route"))
-                            LabeledContent(localized("Output UID"), value: snapshot.currentOutputUID.isEmpty ? localized("Unavailable") : snapshot.currentOutputUID)
-                            LabeledContent(localized("Transport"), value: diagnostics.route.transport)
-                            LabeledContent(localized("Observed / Active Rate"), value: observedAndActiveRateLabel)
-                            LabeledContent(localized("Processing Rate"), value: processingRateLabel)
-                            LabeledContent(localized("Native Output Stream"), value: nativeOutputStreamLabel)
-                            LabeledContent(localized("Physical Output Streams"), value: streamChannelCountsLabel(diagnostics.route.physicalOutputStreamChannelCounts))
-                            LabeledContent(localized("Aggregate Input / Output Streams"), value: aggregateStreamCountsLabel)
-                            LabeledContent(localized("Physical / Aggregate Buffer"), value: routeBufferSizesLabel)
-                            LabeledContent(localized("Physical Safety Offsets in / out"), value: physicalSafetyOffsetsLabel)
-                            LabeledContent(localized("Aggregate Safety Offsets in / out"), value: aggregateSafetyOffsetsLabel)
-                            LabeledContent(
-                                localized("Sample Rate Conversion"),
-                                value: snapshot.metrics.playbackSampleRateConversionActive
-                                    ? localized("Active")
-                                    : localized("Inactive")
-                            )
-
-                            Text(localized("Failure categories can overlap. Timing percentiles use bounded realtime histograms and publish every 1,024 callbacks."))
-                                .font(.caption)
-                                .foregroundStyle(.secondary)
-                                .fixedSize(horizontal: false, vertical: true)
-
-                            Button(localized("Reset metrics")) {
-                                onResetDiagnostics()
-                            }
-                            .controlSize(.large)
-                        }
-                        .padding(.top, 8)
+                VStack(alignment: .leading, spacing: 10) {
+                    Text(localized("Stats for Nerds"))
+                        .font(.headline)
+                    Text(localized("Render timing percentiles, reliability counters, recovery history, and the Core Audio route behind this output."))
+                        .font(.caption)
+                        .foregroundStyle(.secondary)
+                        .fixedSize(horizontal: false, vertical: true)
+                    Button {
+                        isShowingDiagnostics = true
                     } label: {
-                        Text(localized("Stats for nerds"))
-                            .font(.headline)
+                        Label(localized("Show Stats"), systemImage: "waveform.path.ecg.rectangle")
                     }
+                    .controlSize(.large)
+                    .accessibilityHint(Text(localized("Opens detailed audio engine diagnostics")))
                 }
                 .frame(maxWidth: .infinity, alignment: .topLeading)
                 .cardPanel(padding: 16)
+                .sheet(isPresented: $isShowingDiagnostics) {
+                    OutputDiagnosticsSheet(
+                        report: report,
+                        onReset: onResetDiagnostics
+                    )
+                }
             }
             .frame(maxWidth: .infinity, alignment: .topLeading)
         }
@@ -3198,472 +3129,8 @@ private struct OutputTab: View {
         )
     }
 
-    private var addedLatencyLabel: String {
-        if usesSeparateClockDiagnostics {
-            guard snapshot.metrics.playbackBufferObservations > 0 else {
-                return snapshot.isRunning ? localized("Measuring...") : localized("Unavailable")
-            }
-            return bridgeLatencyLabel
-        }
-        guard snapshot.metrics.tapToOutputLatencyObservations > 0 else {
-            return snapshot.isRunning ? localized("Measuring...") : localized("Unavailable")
-        }
-        return tapToOutputLatencyLabel
-    }
-
-    private func diagnosticSectionTitle(_ title: String) -> some View {
-        Text(title)
-            .font(.subheadline.weight(.semibold))
-    }
-
-    @ViewBuilder
-    private var timingRows: some View {
-        let timing = snapshot.metrics.renderTiming
-        LabeledContent(
-            localized("Callback Start Lateness"),
-            value: durationPercentilesLabel(
-                observations: timing.callbackStartLatenessObservations,
-                p50: timing.callbackStartLatenessP50Nanoseconds,
-                p99: timing.callbackStartLatenessP99Nanoseconds,
-                p999: timing.callbackStartLatenessP999Nanoseconds,
-                p9999: timing.callbackStartLatenessP9999Nanoseconds,
-                maximum: timing.maximumCallbackStartLatenessNanoseconds
-            )
-        )
-        if timing.directHeadObservations > 0 {
-            LabeledContent(
-                localized("FIR Head"),
-                value: durationPercentilesLabel(
-                    observations: timing.directHeadObservations,
-                    p50: timing.directHeadP50Nanoseconds,
-                    p99: timing.directHeadP99Nanoseconds,
-                    p999: timing.directHeadP999Nanoseconds,
-                    p9999: timing.directHeadP9999Nanoseconds,
-                    maximum: timing.maximumDirectHeadNanoseconds
-                )
-            )
-            LabeledContent(
-                localized("FIR Tail"),
-                value: durationPercentilesLabel(
-                    observations: timing.tailWorkObservations,
-                    p50: timing.tailWorkP50Nanoseconds,
-                    p99: timing.tailWorkP99Nanoseconds,
-                    p999: timing.tailWorkP999Nanoseconds,
-                    p9999: timing.tailWorkP9999Nanoseconds,
-                    maximum: timing.maximumTailWorkNanoseconds
-                )
-            )
-            LabeledContent(localized("Tail Slack Minimum / Misses"), value: tailCompletionSlackLabel)
-            LabeledContent(
-                localized("FIR Partition Misses"),
-                value: localizedInteger(timing.tailDeadlineMisses)
-            )
-        }
-        LabeledContent(
-            localized("Total Render"),
-            value: durationPercentilesLabel(
-                observations: timing.totalRenderObservations,
-                p50: timing.totalRenderP50Nanoseconds,
-                p99: timing.totalRenderP99Nanoseconds,
-                p999: timing.totalRenderP999Nanoseconds,
-                p9999: timing.totalRenderP9999Nanoseconds,
-                maximum: timing.maximumTotalRenderNanoseconds
-            )
-        )
-        LabeledContent(
-            localized("Completion Lateness"),
-            value: durationPercentilesLabel(
-                observations: timing.completionLatenessObservations,
-                p50: timing.completionLatenessP50Nanoseconds,
-                p99: timing.completionLatenessP99Nanoseconds,
-                p999: timing.completionLatenessP999Nanoseconds,
-                p9999: timing.completionLatenessP9999Nanoseconds,
-                maximum: timing.maximumCompletionLatenessNanoseconds
-            )
-        )
-    }
-
-    private func durationPercentilesLabel(
-        observations: UInt64,
-        p50: UInt64,
-        p99: UInt64,
-        p999: UInt64,
-        p9999: UInt64,
-        maximum: UInt64
-    ) -> String {
-        guard observations > 0 else {
-            return localized("No samples")
-        }
-        let values = [
-            durationPercentileValue(p50, observations: observations, minimum: 2),
-            durationPercentileValue(p99, observations: observations, minimum: 100),
-            durationPercentileValue(p999, observations: observations, minimum: 1_000),
-            durationPercentileValue(p9999, observations: observations, minimum: 10_000),
-            durationPercentileValue(maximum, observations: observations, minimum: 1)
-        ]
-        return localized("\(values.joined(separator: " / ")) us")
-    }
-
-    private func durationPercentileValue(
-        _ nanoseconds: UInt64,
-        observations: UInt64,
-        minimum: UInt64
-    ) -> String {
-        guard observations >= minimum else {
-            return "–"
-        }
-        return localizedDecimal(
-            Double(nanoseconds) / 1_000,
-            minimumFractionDigits: 2,
-            maximumFractionDigits: 2
-        )
-    }
-
-    private var tapToOutputLatencyLabel: String {
-        guard snapshot.metrics.tapToOutputLatencyObservations > 0 else {
-            return localized("No samples")
-        }
-        return localizedLatency(
-            milliseconds: snapshot.metrics.averageTapToOutputLatencyNanoseconds / 1_000_000
-        )
-    }
-
-    private var tapToOutputLatencyRangeLabel: String {
-        guard snapshot.metrics.tapToOutputLatencyObservations > 0 else {
-            return localized("No samples")
-        }
-        let minimum = localizedLatency(
-            milliseconds: Double(snapshot.metrics.minimumTapToOutputLatencyNanoseconds) / 1_000_000
-        )
-        let maximum = localizedLatency(
-            milliseconds: Double(snapshot.metrics.maximumTapToOutputLatencyNanoseconds) / 1_000_000
-        )
-        return localized("\(minimum) to \(maximum)")
-    }
-
-    private var usesSeparateClockDiagnostics: Bool {
-        diagnostics.status.routeMode == .compatibility
-            || diagnostics.status.routeMode == .headsetCompatibility
-    }
-
-    private var tailCompletionSlackLabel: String {
-        let timing = snapshot.metrics.renderTiming
-        guard timing.tailCompletionObservations > 0 else {
-            return localized("No completed blocks")
-        }
-        return localized(
-            "\(localizedInteger(timing.minimumTailCompletionSlackFrames)) frames, \(localizedInteger(timing.tailDeadlineMisses)) misses"
-        )
-    }
-
-    private var diagnosticsResetLabel: String {
-        diagnostics.observation.resetAt?.formatted(date: .abbreviated, time: .standard)
-            ?? localized("Unknown")
-    }
-
-    private var observationDurationLabel: String {
-        diagnosticDurationLabel(diagnostics.observation.observationDurationSeconds)
-    }
-
-    private var runtimeDurationLabel: String {
-        guard let startedAt = diagnostics.observation.runtimeStartedAt else {
-            return localized("Not running")
-        }
-        let duration = diagnosticDurationLabel(
-            diagnostics.observation.runtimeDurationSeconds
-        )
-        return localized(
-            "\(duration), since \(startedAt.formatted(date: .omitted, time: .standard))"
-        )
-    }
-
-    private func diagnosticDurationLabel(_ seconds: Double) -> String {
-        let total = max(Int(seconds.rounded(.down)), 0)
-        if total < 60 {
-            return localized("\(total) seconds")
-        }
-        if total < 3_600 {
-            return localized("\(total / 60) min \(total % 60) sec")
-        }
-        return localized(
-            "\(total / 3_600) hr \((total % 3_600) / 60) min \(total % 60) sec"
-        )
-    }
-
-    private var underrunDetailLabel: String {
-        localized(
-            "\(localizedInteger(snapshot.metrics.playbackUnderrunEvents)) / \(localizedInteger(snapshot.metrics.playbackUnderrunFrames))"
-        )
-    }
-
-    private var droppedFramesLabel: String {
-        localized(
-            "\(localizedInteger(snapshot.metrics.droppedInputFrames)) / \(localizedInteger(snapshot.metrics.droppedBufferedFrames))"
-        )
-    }
-
-    private var deadlineMissesLabel: String {
-        localized(
-            "\(localizedInteger(snapshot.metrics.renderDeadlineMisses)) total, \(localizedInteger(snapshot.metrics.callbackStartStarvations)) start, \(localizedInteger(snapshot.metrics.renderOverruns)) render"
-        )
-    }
-
-    private var discontinuityLabel: String {
-        localized(
-            "\(localizedInteger(snapshot.metrics.inputTimestampDiscontinuities)) input, \(localizedInteger(snapshot.metrics.outputTimestampDiscontinuities)) output, \(localizedInteger(snapshot.metrics.pairedTimestampDiscontinuities)) paired, \(localizedInteger(snapshot.metrics.qualifyingPairedTimestampDiscontinuities)) qualifying"
-        )
-    }
-
-    private var bufferedFramesLabel: String {
-        localized(
-            "\(localizedInteger(snapshot.metrics.currentBufferedFrames)) / \(localizedInteger(snapshot.metrics.maximumPlaybackBufferedFrames)) frames"
-        )
-    }
-
-    private var lastRecoveryLabel: String {
-        guard let reason = diagnostics.recovery.lastReason,
-              let date = diagnostics.recovery.lastRecoveryAt else {
-            return localized("None")
-        }
-        return localized(
-            "\(recoveryReasonLabel(reason)), \(date.formatted(date: .omitted, time: .standard))"
-        )
-    }
-
-    private func recoveryReasonLabel(_ reason: SettingsAudioRecoveryReason) -> String {
-        switch reason {
-        case .renderStall:
-            localized("render stall")
-        case .deadlineMisses:
-            localized("deadline misses")
-        case .timestampDiscontinuity:
-            localized("timestamp discontinuity")
-        case .headsetInstability:
-            localized("headset instability")
-        case .playbackUnderrun:
-            localized("playback underrun")
-        case .adaptiveRenderFailure:
-            localized("adaptive render failure")
-        }
-    }
-
-    private func callbackSizeHistogramLabel(
-        _ observations: [SettingsAudioCallbackSizeObservationDTO]
-    ) -> String {
-        let values = observations.compactMap { observation -> String? in
-            guard observation.observations > 0 else {
-                return nil
-            }
-            let frameSize = observation.frameCount.map(localizedInteger)
-                ?? localized("Other")
-            return "\(frameSize): \(localizedInteger(observation.observations))"
-        }
-        return values.isEmpty ? localized("No samples") : values.joined(separator: ", ")
-    }
-
-    private var callbackPeaksLabel: String {
-        localized(
-            "\(localizedInteger(snapshot.metrics.maximumCaptureCallbackFrames)) / \(localizedInteger(snapshot.metrics.maximumPlaybackCallbackFrames)) frames"
-        )
-    }
-
-    private var timestampJumpLabel: String {
-        guard snapshot.metrics.inputTimestampDiscontinuities > 0
-                || snapshot.metrics.outputTimestampDiscontinuities > 0 else {
-            return localized("No samples")
-        }
-        let input = localizedDecimal(
-            snapshot.metrics.lastInputTimestampJumpFrames,
-            minimumFractionDigits: 1,
-            maximumFractionDigits: 1,
-            signed: true
-        )
-        let output = localizedDecimal(
-            snapshot.metrics.lastOutputTimestampJumpFrames,
-            minimumFractionDigits: 1,
-            maximumFractionDigits: 1,
-            signed: true
-        )
-        return localized("\(input) input / \(output) output frames")
-    }
-
-    private var hostIntervalErrorLabel: String {
-        guard snapshot.metrics.inputTimestampDiscontinuities > 0
-                || snapshot.metrics.outputTimestampDiscontinuities > 0 else {
-            return localized("No samples")
-        }
-        let input = localizedDecimal(
-            Double(snapshot.metrics.lastInputHostIntervalErrorNanoseconds) / 1_000,
-            minimumFractionDigits: 2,
-            maximumFractionDigits: 2,
-            signed: true
-        )
-        let output = localizedDecimal(
-            Double(snapshot.metrics.lastOutputHostIntervalErrorNanoseconds) / 1_000,
-            minimumFractionDigits: 2,
-            maximumFractionDigits: 2,
-            signed: true
-        )
-        return localized("\(input) input / \(output) output us")
-    }
-
-    private var timestampJumpIntervalLabel: String {
-        guard snapshot.metrics.timestampJumpIntervalObservations > 0 else {
-            return localized("No samples")
-        }
-        return minAverageMaxMilliseconds(
-            minimum: snapshot.metrics.minimumTimestampJumpIntervalNanoseconds,
-            average: snapshot.metrics.averageTimestampJumpIntervalNanoseconds,
-            maximum: snapshot.metrics.maximumTimestampJumpIntervalNanoseconds
-        )
-    }
-
-    private var inputAgeLabel: String {
-        guard snapshot.metrics.callbackTimingObservations > 0 else {
-            return localized("No samples")
-        }
-        return minAverageMaxMilliseconds(
-            minimum: snapshot.metrics.minimumInputAgeNanoseconds,
-            average: snapshot.metrics.averageInputAgeNanoseconds,
-            maximum: snapshot.metrics.maximumInputAgeNanoseconds
-        )
-    }
-
-    private var outputLeadLabel: String {
-        guard snapshot.metrics.callbackTimingObservations > 0 else {
-            return localized("No samples")
-        }
-        return minAverageMaxMilliseconds(
-            minimum: snapshot.metrics.minimumOutputLeadNanoseconds,
-            average: snapshot.metrics.averageOutputLeadNanoseconds,
-            maximum: snapshot.metrics.maximumOutputLeadNanoseconds
-        )
-    }
-
-    private func minAverageMaxMilliseconds(
-        minimum: UInt64,
-        average: Double,
-        maximum: UInt64
-    ) -> String {
-        let values = [Double(minimum), average, Double(maximum)].map {
-            localizedDecimal(
-                $0 / 1_000_000,
-                minimumFractionDigits: 2,
-                maximumFractionDigits: 2
-            )
-        }
-        return localized("\(values.joined(separator: " / ")) ms")
-    }
-
-    private var observedAndActiveRateLabel: String {
-        let observed = frequencyOrUnknown(diagnostics.route.observedDeviceSampleRate)
-        let active = frequencyOrUnknown(diagnostics.route.activeDeviceSampleRate)
-        return localized("\(observed) / \(active)")
-    }
-
-    private var processingRateLabel: String {
-        frequencyOrUnknown(diagnostics.route.processingSampleRate)
-    }
-
-    private func frequencyOrUnknown(_ sampleRate: Double) -> String {
-        sampleRate > 0 ? localizedFrequency(sampleRate) : localized("Unknown")
-    }
-
-    private var nativeOutputStreamLabel: String {
-        diagnostics.route.nativeOutputStreamIndex.map {
-            localized("Stream \($0 + 1)")
-        } ?? localized("Unavailable")
-    }
-
-    private func streamChannelCountsLabel(_ counts: [Int]) -> String {
-        guard !counts.isEmpty else {
-            return localized("Unavailable")
-        }
-        return counts.map(localizedInteger).joined(separator: " + ")
-    }
-
-    private var aggregateStreamCountsLabel: String {
-        let input = streamChannelCountsLabel(
-            diagnostics.route.aggregateInputStreamChannelCounts
-        )
-        let output = streamChannelCountsLabel(
-            diagnostics.route.aggregateOutputStreamChannelCounts
-        )
-        return localized("\(input) / \(output)")
-    }
-
-    private var routeBufferSizesLabel: String {
-        let physical = optionalFrameCount(diagnostics.route.physicalDeviceBufferFrameSize)
-        let aggregate = optionalFrameCount(diagnostics.route.aggregateBufferFrameSize)
-        return localized("\(physical) / \(aggregate)")
-    }
-
-    private var physicalSafetyOffsetsLabel: String {
-        safetyOffsetsLabel(
-            input: diagnostics.route.physicalInputSafetyOffsetFrames,
-            output: diagnostics.route.physicalOutputSafetyOffsetFrames
-        )
-    }
-
-    private var aggregateSafetyOffsetsLabel: String {
-        safetyOffsetsLabel(
-            input: diagnostics.route.aggregateInputSafetyOffsetFrames,
-            output: diagnostics.route.aggregateOutputSafetyOffsetFrames
-        )
-    }
-
-    private func safetyOffsetsLabel(input: UInt32?, output: UInt32?) -> String {
-        localized("\(optionalFrameCount(input)) / \(optionalFrameCount(output))")
-    }
-
-    private func optionalFrameCount(_ frames: UInt32?) -> String {
-        frames.map(localizedFrameCount) ?? localized("Unavailable")
-    }
-
-    private var bridgeLatencyLabel: String {
-        localizedLatency(
-            milliseconds: playbackFramesToMilliseconds(
-                snapshot.metrics.averagePlaybackBufferedFrames,
-                bufferSampleRate: snapshot.metrics.playbackBufferSampleRate,
-                fallbackSampleRate: snapshot.currentOutputSampleRate
-            )
-        )
-    }
-
-    private var bridgeLatencyRangeLabel: String {
-        let minimum = playbackFramesToMilliseconds(
-            Double(snapshot.metrics.minimumPlaybackBufferedFrames),
-            bufferSampleRate: snapshot.metrics.playbackBufferSampleRate,
-            fallbackSampleRate: snapshot.currentOutputSampleRate
-        )
-        let maximum = playbackFramesToMilliseconds(
-            Double(snapshot.metrics.maximumPlaybackBufferedFrames),
-            bufferSampleRate: snapshot.metrics.playbackBufferSampleRate,
-            fallbackSampleRate: snapshot.currentOutputSampleRate
-        )
-        return localized(
-            "\(localizedLatency(milliseconds: minimum)) to \(localizedLatency(milliseconds: maximum))"
-        )
-    }
-
-    private var playbackRateCorrectionLabel: String {
-        let correction = localizedDecimal(
-            snapshot.metrics.playbackRateCorrectionPPM,
-            minimumFractionDigits: 1,
-            maximumFractionDigits: 1,
-            signed: true
-        )
-        return localized("\(correction) ppm")
-    }
-
-    private var servoBufferLabel: String {
-        let occupancy = localizedDecimal(
-            snapshot.metrics.filteredPlaybackOccupancyFrames,
-            minimumFractionDigits: 1,
-            maximumFractionDigits: 1
-        )
-        let target = localizedInteger(snapshot.metrics.playbackOccupancyTargetFrames)
-        return localized("\(occupancy) / \(target) frames")
+    private var report: OutputDiagnosticsReport {
+        OutputDiagnosticsReport(snapshot: snapshot)
     }
 
 }
